+++ src/densities.jl
@@ -118,10 +118,6 @@ function ρ_from_total_and_spin(ρtot, ρspin=nothing)
     ρ
 end
 
-################################################################################
-#      densities variations generated by element from the tangent space
-################################################################################
-
 """
 Compute the partial density δρ = Σi φi*conj(δφi) at the indicated ``k``-Point and return it (in Fourier space).
 """
