+++ src/scf/newton.jl
@@ -1,51 +1,293 @@
 # Newton's algorithm for the direct minimization of the energy
+# ------------------------------------------------------------
+#
+# Newton algorithm consist of iterating over density matrices like
+#       P = P - δP
+# where δP solves
+#       (Ω+K)δP = [P,[P,H(P)]] (residual)
+# where (Ω+K) if the Jacobian of the direct minimization iterations. It is
+# defined as a super operator from the space tangent to the constraints
+# manifold at point P_∞, the solution of our problem.
+#   - K is the Hessian of the energy on the manifold;
+#   - Ω represents the influence of the curvature of the manifold :
+#         ΩδP = -[P_∞,[H(P_∞),δP]].
+#     In practice, we dont have access to P_∞ so we just use the current P.
+#
+# Here, in the orbital framework, an element on the tangent space at φ can be written as
+#       δP = Σ |φi><δφi| + hc
+# where the δφi are of size Nb and are all orthogonal to the φj, 1 <= j <= N.
+# Therefore we store them in the same kind of array than φ, with
+# δφ[ik][:,i] = δφi for each k-point.
+# In this framework :
+#   - computing Ωδφ can be done analitically with the formula
+#         Ωδφi = H*δφi - Σj <δφj|H|δφi> δφj - Σj λji δφj
+#     where λij = <φi|H|φj>;
+#   - computing Kδφ can be done using the kernel function of DFTK after
+#     computing the associated δρ;
+#   - to perform Newton iterations, we solve (Ω+K)δφ = Hφ-λφ and then perform
+#     φ = φ - δφ, with proper orthonormalization.
+#
+# For further details :
+# Eric Cancès, Gaspard Kemlin, Antoine Levitt. Convergence analysis of
+# direct minimization and self-consistent iterations. SIAM Journal of Matrix
+# Analysis and Applications, 42(1):243–274 (2021)
+
+import KrylovKit: ArnoldiIterator, Orthogonalizer, OrthonormalBasis, KrylovDefaults, orthogonalize!
+using KrylovKit
+
+############################### TOOLS ##########################################
+
+"""
+    compute_residual(basis::PlaneWaveBasis, φ, occ)
+
+Compute the residual associated to a set of planewave φ, that is to say
+H(φ)*φ - λ*φ where λ is the set of rayleigh coefficients associated to the φ.
+"""
+function compute_residual(basis::PlaneWaveBasis, φ, occ)
+
+    # necessary quantities
+    Nk = length(basis.kpoints)
+    ρ = compute_density(basis, φ, occ)
+    energies, H = energy_hamiltonian(basis, φ, occ; ρ=ρ)
+
+    # compute residual
+    res = similar(φ)
+    for ik = 1:Nk
+        φk = φ[ik]
+        N = size(φk, 2)
+        Hk = H.blocks[ik]
+        # eigenvalues as rayleigh coefficients
+        egvalk = φk'*(Hk*φk)
+        # compute residual at given kpoint as H(φ)φ - λφ
+        res[ik] = Hk*φk - φk*egvalk
+    end
+    res
+end
+
+# To project onto the space tangent to φ, we project δφ onto the orthogonal of φ
+function proj(δφ, φ; tol_test=1e-12)
+
+    Nk1 = size(δφ,1)
+    Nk2 = size(φ,1)
+    @assert Nk1 == Nk2
+    Nk = Nk1
+
+    Πδφ = similar(δφ)
+
+    for ik = 1:Nk
+        φk = φ[ik]
+        δφk = δφ[ik]
+        Πδφk = deepcopy(δφk)
+
+        N1 = size(δφk,2)
+        N2 = size(φk,2)
+        @assert N1 == N2
+        N = N1
+
+        for i = 1:N, j = 1:N
+            Πδφk[:,i] -= (φk[:,j]'δφk[:,i]) * φk[:,j]
+        end
+        Πδφ[ik] = Πδφk
+    end
+
+    # test orthogonalisation
+    for ik = 1:Nk
+        φk = φ[ik]
+        δφk = δφ[ik]
+        Πδφk = Πδφ[ik]
+        N = size(φk,2)
+        for i = 1:N, j = 1:N
+            @assert abs(Πδφk[:,i]'φk[:,j]) < tol_test [println(abs(Πδφk[:,i]'φk[:,j]))]
+        end
+    end
+
+    Πδφ
+end
+
+# packing routines
+pack(φ) = vcat(Base.vec.(φ)...)
+function packing(basis::PlaneWaveBasis{T}, φ) where T
+
+    Nk = length(basis.kpoints)
+
+    lengths = [length(φ[ik]) for ik = 1:Nk]
+    starts = copy(lengths)
+    starts[1] = 1
+    for ik = 1:Nk-1
+        starts[ik+1] = starts[ik] + lengths[ik]
+    end
+    unpack(x) = [@views reshape(x[starts[ik]:starts[ik]+lengths[ik]-1], size(φ[ik]))
+                 for ik = 1:Nk]
+
+    packed_proj(δφ,φ) = proj(unpack(δφ), unpack(φ))
+
+    (pack, unpack, packed_proj)
+end
+
+# KrylovKit custom orthogonaliser to be used in KrylovKit eigsolve, svdsolve,
+# linsolve, ...
+# This has to be passed in KrylovKit solvers when studying (Ω+K) as a super
+# operator from the tangent space to the tangent space to be sure that
+# iterations are constrained to stay on the tangent space.
+struct OrthogonalizeAndProject{F, O <: Orthogonalizer, φ} <: Orthogonalizer
+    projector::F
+    orth::O
+    φ::φ
+end
+OrthogonalizeAndProject(projector, φ) = OrthogonalizeAndProject(projector,
+                                                                KrylovDefaults.orth,
+                                                                φ)
+function KrylovKit.orthogonalize!(v::T, b::OrthonormalBasis{T}, x::AbstractVector,
+                                        alg::OrthogonalizeAndProject) where {T}
+    v, x = orthogonalize!(v, b, x, alg.orth)
+    v = reshape(v, size(alg.φ))
+    v = pack(alg.projector(v, alg.φ))::T
+    v, x
+end
+function KrylovKit.orthogonalize!(v::T, x::AbstractVector,
+                                        alg::OrthogonalizeAndProject) where {T}
+    v, x = orthogonalize!(v, x, alg.orth)
+    v = reshape(v, size(alg.φ))
+    v = pack(alg.projector(v, alg.φ))::T
+    v, x
+end
+function KrylovKit.gklrecurrence(operator, U::OrthonormalBasis, V::OrthonormalBasis, β,
+                                 alg::OrthogonalizeAndProject)
+    u = U[end]
+    v = operator(u, true)
+    v = axpy!(-β, V[end], v)
+    # for q in V # not necessary if we definitely reorthogonalize next step and previous step
+    #     v, = orthogonalize!(v, q, ModifiedGramSchmidt())
+    # end
+    α = norm(v)
+    rmul!(v, inv(α))
+
+    r = operator(v, false)
+    r = axpy!(-α, u, r)
+    for q in U
+        r, = orthogonalize!(r, q, alg)
+    end
+    β = norm(r)
+    return v, r, α, β
+end
+
+############################# OPERATORS ########################################
+
+"""
+    apply_Ω(basis::PlaneWaveBasis, δφ, φ, H, λ)
+
+Compute the application of Ω to an element in the space tangent to φ.
+It can be done analitically with the formula
+      Ωδφi = H*δφi - Σj <δφj|H|δφi> δφj - Σj λji δφj
+where λij = <φi|H|φj>;
+"""
+function apply_Ω(basis::PlaneWaveBasis, δφ, φ, H)
+    Nk = length(basis.kpoints)
+    Ωδφ = similar(φ)
+
+    for ik = 1:Nk
+        φk = φ[ik]
+        δφk = δφ[ik]
+        Hk = H.blocks[ik]
+        λk = φk'*(Hk*φk)
+        Ωδφk = similar(δφk)
+
+        N1 = size(δφk,2)
+        N2 = size(φk,2)
+        @assert N1 == N2
+        N = N1
+
+        for i = 1:N
+            Hδφki = Hk * δφk[:,i]
+            Ωδφk[:,i] = Hδφki
+            for j = 1:N
+                Ωδφk[:,i] -= (φk[:,j]'Hδφki) * φk[:,j]
+                Ωδφk[:,i] -= λk[j,i] * δφk[:,j]
+            end
+        end
+        Ωδφ[ik] = Ωδφk
+    end
+    # ensure proper projection onto the tangent space
+    proj(Ωδφ, φ)
+end
+
+"""
+    apply_K(basis::PlaneWaveBasis, δφ, φ, ρ, occ)
+
+Compute the application of K to an element in the space tangent to φ
+with the same notations as for the computation of Ω, we have
+      Kδφi = Π(dV * φi),
+Π being the projection on the space tangent to the φ and
+      dV = kernel(δρ)
+where δρ = Σi φi*conj(δφi) + hc.
+"""
+function apply_K(basis::PlaneWaveBasis, δφ, φ, ρ, occ)
+    Nk = length(basis.kpoints)
+
+    δρ = compute_density(basis, φ, δφ, occ)
+    dV = apply_kernel(basis, δρ; ρ=ρ)
+    Kδφ = similar(φ)
+
+    for ik = 1:Nk
+        kpt = basis.kpoints[ik]
+        φk = φ[ik]
+        dVφk = similar(φk)
+
+        for i = 1:size(φk,2)
+            φk_r = G_to_r(basis, kpt, φk[:,i])
+            dVφk_r = dV[:,:,:,kpt.spin] .* φk_r
+            dVφk[:,i] = r_to_G(basis, kpt, dVφk_r)
+        end
+        Kδφ[ik] = dVφk
+    end
+    # ensure proper projection onto the tangent space
+    proj(Kδφ, φ)
+end
+
+############################# NEWTON ALGORITHM #################################
 
 """
     newton_step(basis::PlaneWaveBasis, φ, φproj, res, occ;
-                tol_krylov=1e-12, krylov_verb=1)
+                tol_krylov=1e-12, krylov_verbosity=1)
 
 Perform a newton step : we take as given a planewave set φ and we return the
 newton step δφ and the updated state φ - δφ
 (after proper orthonormalization) where δφ solves Jac * δφ = res
 and Jac is the Jacobian of the projected gradient descent : Jac = Ω+K.
-(cf. Eric Cancès, Gaspard Kemlin, Antoine Levitt. Convergence analysis of
-direct minimization and self-consistent iterations. SIAM Journal of Matrix
-Analysis and Applications, 42(1):243–274 (2021))
+δφ is an element of the tangent space at φproj (set to φ if not specified in newton function).
 """
 function newton_step(basis::PlaneWaveBasis, φ, φproj, res, occ;
-                     tol_krylov=1e-12, krylov_verb=1)
+                     tol_krylov=1e-12, krylov_verbosity=1)
 
     # necessary quantities
     N = size(φ[1],2)
     Nk = length(basis.kpoints)
-    ortho(ψk) = Matrix(qr(ψk).Q)
+    ortho(φk) = Matrix(qr(φk).Q)
 
     # compute quantites at the point which define the tangent space
     ρproj = compute_density(basis, φproj, occ)
     energies, Hproj = energy_hamiltonian(basis, φproj, occ; ρ=ρproj)
-    λproj = similar(φ)
-    for ik = 1:Nk
-        φk = φ[ik]
-        Hk = Hproj.blocks[ik]
-        λproj[ik] = φk'*(Hk*φk)
-    end
 
     # packing routines
     pack, unpack, packed_proj = packing(basis, φproj)
 
-    # solve linear system with KrylovKit
+    # mapping of the linear system on the tangent space
     function f(x)
         δφ = unpack(x)
-        ΩpKx = ΩplusK(basis, δφ, φproj, ρproj, Hproj, λproj, occ)
-        pack(ΩpKx)
+        δφ = proj(δφ, φproj)
+        Kδφ = apply_K(basis, δφ, φproj, ρproj, occ)
+        Ωδφ = apply_Ω(basis, δφ, φproj, Hproj)
+        pack(Ωδφ .+ Kδφ)
     end
 
     # project res on the good tangent space before starting
     res = proj(res, φproj)
 
-    # solve (Ω+K) δφ = res
+    # solve (Ω+K) δφ = res on the tangent space with KrylovKit
     δφ, info = linsolve(f, pack(res);
-                        tol=tol_krylov, verbosity=krylov_verb,
+                        tol=tol_krylov, verbosity=krylov_verbosity,
+                        # important to specify custom orthogonaliser to keep
+                        # iterations on the tangent space
                         orth=OrthogonalizeAndProject(packed_proj, pack(φproj)))
     δφ = unpack(δφ)
     # ensure proper projection onto the tangent space
@@ -61,13 +303,15 @@ end
 
 """
     newton(basis::PlaneWaveBasis; ψ0=nothing,
-           tol=1e-6, max_iter=100, φproj=nothing, verb=1)
+           tol=1e-6, max_iter=100, φproj=nothing, verbosity=0)
 
 Newton algorithm. Be careful that the starting needs to be not too far from the solution.
 Still contains debugging features.
+If φproj is nothing, we use the successive φ's to define the tangent spaces on which we solve
+Newton's equation. φproj can be useful when we want to fix the tangent space for testing.
 """
 function newton(basis::PlaneWaveBasis{T}; ψ0=nothing,
-                tol=1e-6, max_iter=100, φproj=nothing, verb=0) where T
+                tol=1e-6, max_iter=20, φproj=nothing, verbosity=0) where T
 
     ## setting parameters
     model = basis.model
@@ -80,9 +324,9 @@ function newton(basis::PlaneWaveBasis{T}; ψ0=nothing,
     Nk = length(basis.kpoints)
     occupation = [filled_occ * ones(T, N) for ik = 1:Nk]
 
-    ## starting point
+    ## starting point and keep only occupied orbitals
     if ψ0 === nothing
-        ortho(ψk) = Matrix(qr(ψk).Q)
+        ortho(φk) = Matrix(qr(φk).Q)
         ψ0 = [ortho(randn(Complex{T}, length(G_vectors(kpt)), N))
               for kpt in basis.kpoints]
     else
@@ -90,7 +334,6 @@ function newton(basis::PlaneWaveBasis{T}; ψ0=nothing,
             ψ0[ik] = ψ0[ik][:,1:N]
         end
     end
-    ## keep only occupied orbitals
     if φproj != nothing
         for ik in 1:Nk
             φproj[ik] = φproj[ik][:,1:N]
@@ -123,7 +366,7 @@ function newton(basis::PlaneWaveBasis{T}; ψ0=nothing,
         # compute next step
         res = compute_residual(basis, φ, occupation)
         φ, δφ = newton_step(basis, φ, ϕ, res, occupation;
-                            krylov_verb=verb)
+                            krylov_verbosity=verbosity)
 
         # compute error on the densities and the energies
         ρ_next = compute_density(basis, φ, occupation)
@@ -140,9 +383,8 @@ function newton(basis::PlaneWaveBasis{T}; ψ0=nothing,
         prev_energies = energies
     end
 
-    energies, H = energy_hamiltonian(basis, φ, occupation; ρ=ρ)
-
     # Rayleigh-Ritz
+    energies, H = energy_hamiltonian(basis, φ, occupation; ρ=ρ)
     eigenvalues = []
     for ik = 1:Nk
         Hφk = H.blocks[ik] * φ[ik]
@@ -151,10 +393,8 @@ function newton(basis::PlaneWaveBasis{T}; ψ0=nothing,
         φ[ik] .= φ[ik] * F.vectors
     end
 
-
     εF = nothing  # does not necessarily make sense here, as the
                   # Aufbau property might not even be true
-
 
     # return results
     (ham=H, basis=basis, energies=energies, converged=true,
